fix(db,test): FK type mismatch on product_id + Pint failure in tests/Pest.php
CI was failing on MySQL 8 because every migration that references
products.id declared product_id as BIGINT UNSIGNED while products.id is
INT UNSIGNED (legacy $table->increments('id')). MySQL 8 enforces strict
FK type compatibility and aborted the installer on the first sibling
migration (Amazon, alphabetically). SQLite — used in local test runs —
silently accepted the mismatch, which is why this didn't surface locally.

Fixed 12 migrations to declare product_id as unsignedInteger:

  • Amazon, Ebay, Magento2, Noon, Salla, EasyOrders, WooCommerce mapping
    tables (7): swapped foreignId('product_id')->constrained('products')
    for explicit unsignedInteger + foreign()->references()->on() pair.
  • ChannelConnector: product_channel_mappings, channel_sync_conflicts
  • Pricing: product_costs, margin_protection_events
  • EasyOrders: unified_order_items

When products is eventually migrated to bigIncrements, all 12 of these
column declarations will need to flip to unsignedBigInteger in lock-step;
the comment in each file flags the dependency.

Also extracted the FQCN uses() calls in tests/Pest.php into explicit
`use` statements at the top so Pint's preset (no fully-qualified names
inside uses()) passes — that single file was the entire Linting Tests
failure. (130+ other pre-existing Pint violations across the repo will
be cleaned up in a follow-up blanket pint commit.)

+++ packages/Webkul/ChannelConnector/src/Database/Migrations/2026_02_14_100004_create_product_channel_mappings_table.php
@@ -13,7 +13,8 @@ return new class extends Migration
             $table->id();
             $table->unsignedBigInteger('tenant_id');
             $table->unsignedBigInteger('channel_connector_id');
-            $table->unsignedBigInteger('product_id');
+            // products.id is INT UNSIGNED (legacy $table->increments('id')); FK column must match.
+            $table->unsignedInteger('product_id');
             $table->string('external_id');
             $table->string('external_variant_id')->nullable();
             $table->string('entity_type', 20)->default('product');
